Refactored the suggestion service to be interface-based
I wanted to put some restrictions around the SuggestionService
and how it works and I didn't want the fact that we're using
OpenAI's LLM to dominate the interface for the client. As
such, I built a SuggestionService interface then implemented
an LLMSuggestionService that implements it. I also separated
out an LLMClient that the LLMSuggestionService calls.