Dockerize app
In this commit, I'm adding a Dockerfile that builds the
production code and creates an Nginx container to serve
the files.

Along the way, I had to upgrade Node to a higher version.
I believe it was React why I needed to upgrade it. I
don't quite recall.

Finally, I'm adding a Github workflow to push the
image to Docker Hub. It will be used by the Home Hub.